hrtimer: Undo half of 190d14f
I discarded the hrtimer.c part of commit 190d14f ("hrtimer:
Annotate lockless access to timer->state") because it conflicted
with commit 6894406 ("hrtimer: update timer->state with
'pinned' information") and I wanted to be cautious. Unfortunately,
I did not realize there was an hrtimer.h modification as well; undo
that so there is no mismatch.

Reported-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>